+++ webroot/templates/index.html
@@ -9,13 +9,13 @@
 <body>
     <nav class="nav-menu">
         <ul>
-            <li><a href="{{ url_for('home') }}">Home</a></li>
-            <li><a href="{{ url_for('home') }}">About US</a></li>
-            <li><a href="{{ url_for('home') }}">Flower Bouquets</a></li>
-            <li><a href="{{ url_for('home') }}">Gourmet Goods</a></li>
-            <li><a href="{{ url_for('home') }}">Gift Baskets</a></li>
-            <li><a href="{{ url_for('home') }}">Updates</a></li>
-            <li><a href="{{ url_for('home') }}">Account</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Home</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">About US</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Flower Bouquets</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Gourmet Goods</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Gift Baskets</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Updates</a></li>
+            <li><a href="{{ url_for('webpages.home') }}">Account</a></li>
         </ul>
     </nav>
